// src/types.ts
export interface Ticket {
  origin: string;
  origin_name: string;
  destination: string;
  destination_name: string;
  departure_date: string;
  departure_time: string;
  arrival_date: string;
  arrival_time: string;
  carrier: string;
  stops: number;
  price: number;
}

export enum CurrencyEnum {
  RUB = "RUB",
  USD = "USD",
  EUR = "EUR",
}